+++ src/graphql/schema.graphql
@@ -1,0 +1,12 @@
+type Bus {
+    connected: Boolean!
+    connecting: Boolean!
+    devices: [Device!]!
+    device(deviceId: String): Device
+}
+type Device {
+    deviceId: ID
+    shortId: String!
+    name: String!
+    services: [Int!]!
+}
